Refactor and add tests.
This revision changed the names of both the function and the
parameters. It also adds a simple test that you can use to
verify the operation of variants. Please add more test cases
at will.

The test has been verified as passing. There are ways to stop
changes that brea tsts from being checked in, but we'll get to
that later.